lots of refactoring for registration
