+++ packages/unlock-js/src/abis.ts
@@ -1217,7 +1217,7 @@ const abis = {
         'function withdraw(address _tokenAddress,address _recipient,uint256 _amount)',
       ],
       bytecodeHash:
-        '0xafa54abfa6cc8380d911e238c3f7e198cf12180e68ee8cdc7d8f2d8e87e5569b',
+        '0x349cbd03803a9577581041659dfa1d832ca72d580bdb998e38907e98f20ad649',
     },
   },
   Unlock: {
@@ -1632,7 +1632,7 @@ const abis = {
         'function weth() view returns (address)',
       ],
       bytecodeHash:
-        '0x4732ddf02736b94eeeccfb2ae3ddea4295878ace8bfe637b6dbb69a55c341ee5',
+        '0x091d4f41845ea7a4880d62d7068ea5c85128b9ed2879d7ae0402ae7b6b2d3f8f',
     },
   },
 }
