+++ unlock-app/src/components/content/event/Registration/HasTicket.tsx
@@ -1,0 +1,13 @@
+import Link from 'next/link'
+
+export const HasTicket = () => {
+  return (
+    <p className="text-lg">
+      🎉 You already have a ticket! You can view it in{' '}
+      <Link className="underline" href="/keychain">
+        your keychain
+      </Link>
+      .
+    </p>
+  )
+}
